ui: Remove save buttons in workbench panel + add save on blur (aggressive) + disable validate on save
